refactor: move the WhatsApp provider cache onto a mixin
The number, template and refresh caches only apply to Meta Cloud API providers,
so they now live in WhatsAppProviderMixin rather than swelling MessagingProvider.
